+++ src/index.ts
@@ -2,6 +2,14 @@ import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js";
 import { z } from "zod";
 import fetch from "node-fetch";
+import fs from "node:fs";
+import path from "node:path";
+import type {
+  HaikuMonument,
+  HaikuMonumentResponse,
+  GeoJSONFeatureCollection,
+  GeoJSONFeature,
+} from "./types.js";
 
 const server = new McpServer({
   name: "kuhi-api",
@@ -10,53 +18,13 @@ const server = new McpServer({
 
 const API_BASE_URL = "https://api.kuhiapi.com";
 
-interface HaikuMonument {
-  id: number;
-  text: string;
-  established_date: string;
-  commentary: string;
-  image_url: string;
-  created_at: string;
-  updated_at: string;
-  poet_id: number;
-  source_id: number;
-  location_id: number;
-  poets: Array<{
-    id: number;
-    name: string;
-    biography: string;
-    links: string;
-    image_url: string;
-    created_at: string;
-    updated_at: string;
-  }>;
-  sources: Array<{
-    id: number;
-    title: string;
-    author: string;
-    year: number;
-    url: string;
-    publisher: string;
-    created_at: string;
-    updated_at: string;
-  }>;
-  locations: Array<{
-    id: number;
-    prefecture: string;
-    region: string;
-    address: string;
-    latitude: number;
-    longitude: number;
-    name: string;
-  }>;
-}
-
-async function fetchHaikuMonuments() {
+async function fetchHaikuMonuments(): Promise<HaikuMonument[]> {
   const response = await fetch(`${API_BASE_URL}/haiku-monuments`);
   if (!response.ok) {
     throw new Error("Failed to fetch haiku monuments");
   }
-  return response.json() as Promise<HaikuMonument[]>;
+  const data = (await response.json()) as HaikuMonumentResponse;
+  return data.haiku_monuments;
 }
 
 async function fetchHaikuMonumentsByRegion(region: string) {
@@ -92,6 +60,54 @@ async function fetchHaikuMonumentsByCoordinates(
     throw new Error("Failed to fetch haiku monuments by coordinates");
   }
   return response.json() as Promise<HaikuMonument[]>;
+}
+
+function convertToGeoJSON(monuments: HaikuMonument[]): GeoJSONFeatureCollection {
+  const features: GeoJSONFeature[] = monuments.map((monument) => {
+    const location = monument.locations[0];
+    return {
+      type: "Feature",
+      geometry: {
+        type: "Point",
+        coordinates: [location.longitude, location.latitude] as [number, number],
+      },
+      properties: {
+        id: monument.id,
+        text: monument.text,
+        established_date: monument.established_date,
+        commentary: monument.commentary,
+        image_url: monument.image_url,
+        poet_name: monument.poets[0]?.name || "",
+        prefecture: location.prefecture,
+        region: location.region,
+        address: location.address,
+        name: location.name,
+      },
+    };
+  });
+
+  return {
+    type: "FeatureCollection",
+    features,
+  };
+}
+
+async function generateGeoJSONFile(outputPath: string): Promise<void> {
+  try {
+    const monuments = await fetchHaikuMonuments();
+    const geojson = convertToGeoJSON(monuments);
+    
+    const outputDir = path.dirname(outputPath);
+    if (!fs.existsSync(outputDir)) {
+      fs.mkdirSync(outputDir, { recursive: true });
+    }
+    
+    fs.writeFileSync(outputPath, JSON.stringify(geojson, null, 2));
+    console.log(`GeoJSONファイルを生成しました: ${outputPath}`);
+  } catch (error) {
+    console.error("GeoJSONファイルの生成中にエラーが発生しました:", error);
+    throw error;
+  }
 }
 
 server.tool(
@@ -138,9 +154,28 @@ server.tool(
   },
 );
 
+server.tool(
+  "get_haiku_monuments_geojson",
+  "句碑データベースに登録されているすべての句碑の情報をGeoJSON形式で表示",
+  {},
+  async () => {
+    const monuments = await fetchHaikuMonuments();
+    const geojson = convertToGeoJSON(monuments);
+    return { content: [{ type: "text", text: JSON.stringify(geojson) }] };
+  },
+);
+
 async function main() {
   const transport = new StdioServerTransport();
   await server.connect(transport);
+  
+  if (process.argv[2]) {
+    try {
+      await generateGeoJSONFile(process.argv[2]);
+    } catch (error) {
+      console.error("GeoJSONファイルの生成に失敗しました:", error);
+    }
+  }
 }
 
 main().catch((error) => {
