+++ tsconfig.json
@@ -11,6 +11,6 @@
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true
   },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist", ".git"]
+  "include": ["src/index.ts"],
+  "exclude": ["node_modules", "dist", ".git", "src/types.ts"]
 }
